deluser/requirements.txt
requests
Flask==2.0.2
google-python-cloud-debugger
google-cloud-logging
google-cloud-tasks==2.7.1
google-cloud-firestore